requirements.txt
pillow~=11.0.0
torch~=2.5.1
torchvision~=0.20.1
tqdm~=4.66.6